pyproject.toml
[project]
name = "piper-kit"
version = "0.1.0"
dependencies = [
    "python-can>=4.5.0",
]
requires-python = ">=3.13"
authors = [
    {name = "Alfi Maulana", email = "alfi.maulana.f@gmail.com"},
]
description = "SDK and CLI tools for AgileX PiPER robotic arm"
readme = "README.md"
license = "MIT"
keywords = ["sdk", "cli", "robotic", "arm"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3 :: Only",
]

[project.scripts]
piper = "piper_kit._cli.main:main"

[project.urls]
repository = "https://github.com/threeal/piper-kit.git"
issues = "https://github.com/threeal/piper-kit/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "lefthook>=1.11.14",
    "ruff>=0.11.13",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812"]

[tool.ruff.lint.per-file-ignores]
"__main__.py" = ["T201"]
